Fix SA1200 UsingDirectivesMustBePlacedWithinNamespace
