+++ DNN Platform/Library/Entities/Content/Common/ContentExtensions.cs
@@ -1,36 +1,18 @@
 ﻿// Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information
-using System;
-using System.Collections.Generic;
-using System.Collections.Specialized;
-using System.Linq;
-using System.Text;
-
-using DotNetNuke.Common.Utilities;
-using DotNetNuke.Entities.Content.Taxonomy;
-
-namespace DotNetNuke.Entities.Content
-{
-    [Obsolete("Moving ContentExtensions to the DotNetNuke.Entities.Content namespace was an error. Please use DotNetNuke.Entities.Content.Common.ContentExtensions. Scheduled removal in v10.0.0.")]
-    public static class ContentExtensions
-    {
-        // only forwarding public methods that existed as of 6.1.0
-        // calls to internal methods will be fixed in the source
-        public static string ToDelimittedString(this List<Term> terms, string delimiter)
-        {
-            return Common.ContentExtensions.ToDelimittedString(terms, delimiter);
-        }
-
-        public static string ToDelimittedString(this List<Term> terms, string format, string delimiter)
-        {
-            return Common.ContentExtensions.ToDelimittedString(terms, format, delimiter);
-        }
-    }
-}
 
 namespace DotNetNuke.Entities.Content.Common
 {
+    using System;
+    using System.Collections.Generic;
+    using System.Collections.Specialized;
+    using System.Linq;
+    using System.Text;
+
+    using DotNetNuke.Common.Utilities;
+    using DotNetNuke.Entities.Content.Taxonomy;
+
     /// <summary>
     /// Extension methods for Term, Vocabulary, ContentItem.
     /// </summary>
